pyproject.toml
[project]
name = "testmind-ai"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "aiofiles>=24.1.0",
    "aioredis>=2.0.1",
    "alembic>=1.16.4",
    "asyncpg>=0.30.0",
    "fastapi>=0.116.1",
    "httpx>=0.28.1",
    "langchain>=0.3.26",
    "langchain-openai>=0.3.28",
    "markdown>=3.8.2",
    "pdfplumber>=0.11.7",
    "pydantic>=2.11.7",
    "pydantic-settings>=2.10.1",
    "pypdf>=5.1.0",
    "pytest>=8.4.1",
    "pytest-asyncio>=1.0.0",
    "pytest-cov>=6.2.1",
    "pytest-html>=4.1.1",
    "pytest-json-report>=1.5.0",
    "psutil>=6.1.0",
    "python-docx>=1.2.0",
    "python-frontmatter>=1.1.0",
    "python-multipart>=0.0.20",
    "redis>=6.2.0",
    "sqlalchemy>=2.0.41",
    "tiktoken>=0.9.0",
    "uvicorn>=0.35.0",
]
